fix datatypes and upto

+++ db/migrate/20181126112719_create_days.rb
@@ -2,7 +2,7 @@ class CreateDays < ActiveRecord::Migration[5.2]
   def change
     create_table :days do |t|
       t.string :date
-      t.integer :revenue
+      t.decimal :revenue
       t.references :good, foreign_key: true
 
       t.timestamps
